Adds a separate resource registration recipe
In order to make chef-client runs simpler and faster, we break
out a separate registration recipe that registers users, tenants,
services, endpoints, roles, etc with the Keystone server. This
recipe only needs to be in the run_list of one controller node
instead of all of them, which currently have the server recipe
in their runlists.

Removes the admin password/tenant/username stuff from the
credentials and register Chef providers and just replaces that
round-trip request to Keystone with the secret bootstrap token,
which should speed up the registration requests substantially.