Updated tests with guard support
ChefSpec just released the ability to test guards.  We had tests
which tested functionality incorrectly, due to chefspec not supporting
guards.  Updated quite a few tests to be clearer, and to test
when should and should not occur.  Also, removed some duplicate SuSE
tests in server_spec.rb.  The vendor tests should test differences,
not the exact same things as the main test.

Change-Id: Id9e8c4f6566ffb8910d98ddb6bf19f49412aa69c