Removed the db recipe
The db recipe have been removed. It is up to the operator to
setup the database, users, and grants before deploying openstack.
It has been moved to the `openstack-db` recipe in the
`openstack-ops-database` wrapper cookbook.  Needed to converge
in the register tests, so Chef loads the provider and creates
the classes.  The db spec was previously doing that, which allowed
tests to pass.

Change-Id: I301b7fc625790960efd99d08395b91204ae18107