There is no need to bash boostrap keystone
With the introduction of the keystone LWRP back in the Grizzly days,
this fragile bootstrap routine was made obsolete. The LWRP already
makes use of the bootstrap token to create various new users by way
of the registration recipe and/or a custom recipe defined by the user
which declares keystone LWRP resources.

Also, this is Chef, not bash.

Change-Id: Iac688e28a17c1ec0eb3a57da5f9a0f54b66e503d