keystone-manage to be run by keystone user/group
keystone-manage pki_setup and keystone-manage db_sync should be run by
the designated keystone user and group. The reason being that these
commands when run as root (as is the case with chef) may drop a
keystone.log file. If this is the first pass, that file will be owned
by root and subsequent writes to that file by the proper keystone user
or group will likely fail.

This change also removes a redundant test for opensuse.

Change-Id: I627674b319738f9c79ff146ddfdf979db55b2c14