Add ability to specify IP for service
The problem that this change addresses is that the address_for method
will not work correctly if there are multiple IP address associated
with the specified interface.

The approach to solving this problem and moving towards the overall
goal of having one place where service networking information is
stored is to convert address_for calls into endpoints, and add a
address() method to the endpoints interface for IP address resolution.

The address() method has the following behavior: if the
bind_interface of an endpoint is set, then the IP is looked up on
the interface.  Otherwise, the IP specified in the host attribute is
returned.  This allows the caller to choose either method of
determining what IP a service will be bound to.

This change switches the identity cookbook to use an endpoint for
binding instead of address_for.

This change is dependent on the follow change which adds the endpoint
to common:
https://review.openstack.org/#/c/79387/

blueprint increase-ip-binding-flexibility

Change-Id: Ibccc187c5b04e78468d289d175407007451a4b2b