Add crontab to flush tokens via keystone-manage
When the keystone tokens backend is SQL, the tokens table grows
unboundedly as new tokens are issued and not disposed of after
expiration.

Keystone provides a tool to delete the tokens, and this patch puts it in
a cronjob in order to avoid growing the tokens database forever.

Change-Id: I3e4831619efec273ac5aa378fff4b14ad877f326
Implements: blueprint token-flush