Only manage the keystone-paste.ini if specified
There are no attributes in this template file, which means it just
overrides the existing keystone-paste.ini, and needs to be updated for
each release.

Additionally, the current template file adds in options that are listed
as deprecated, resulting in warnings in keystone.log. Removing the
template file will mean the packaged keystone-paste.ini is used.

This patch removes the keystone-paste.ini template and adds an attribute
to allow a keystone-paste.ini file to be specified.

Change-Id: Id2e6df82acea480320eaf3d07c0570802e4424ca
Closes-Bug: #1313828