Fix the crontab entry for keystone-manage token_flush
I'm confused by the use of

  `which foo` --args

since if which is able to find `foo`, so is the shell, hence

  foo --args

should be equivalent.

Also, echoing the exit code $? when run with && is not interesting, it
will always be 0.  Now, a failure will be reported, too.

Changed the format slightly to get rid of the %Q{} worries.

Change-Id: I39116496747d58fd816201d07713377bb193fc8f