internalURL should use the internalURL variable
This patches what appears to be a typo where the internalURL passed to
openstack_identity_register with the "adminURL" variable instead of the
internalURL variable that is defined earlier in the code.

This was noticed as the cookbook was setting up the endpoints with the
internalURL defined to the admin endpoint.  This is opposite to
OpenStack documentation and breaks operations of other openstack
projects which rely on internalURL.  For example the admin pipeline does
not have user_CRUD, while the public one does, breaking dashboard.

Following the inclusion of this patch the identity-internal endpoint
will need to be added to cookbook-chef-common.  Following that addition
the number of node.set for the identity-internal endpoint can be
reduced.

Partial-Bug: #1356732

Change-Id: I4bc3d49fd15b7254976a5ed8e964a4b3e3af02c6