Allow create_user to find existing user regardless of tenant
The current code was checking for existing user ONLY in the
requested tenant.  But if the user already existing with another
tenant, the create user call will fail. With this patch, it will
skip the create and succeed, and then the grant_role action can
be used to add a role for this user in the requested tenant.

Change-Id: I0a96ed6103e9a7da7f2e50624ddaff45d2ed9881
Closes-Bug: 1457533