Allow keystone under apache
Keystone is recommended to run under apache and the service side
is already deprecated and will be removed in the M release.
This patch adds a new recipe, server-apache, to allow keystone
under apache.

I intentionally just copied the existing server recipe and spec to
create the new ones and I figure those will just be removed in
the M release anyway, no need for "common" type code here.

The majority of the recipe code is exactly the same, just the
last lines, "Start of Apache..." have been added.

This should also work with the existing dashboard cookbook with the
one exception that the apache 3.1 cookbook is needed.  There's already
a patch out for that.

I don't plan on changing the default kilo role to use this, that will
be done once we branch for liberty.

Change-Id: I1641e1e5c6bf56d0765ef6e54ae32848431f6d6e
Implements: blueprint keystone-apache