Remove unecessary calls to keystone cli
I noticed that the create_user and grant_role actions did some
unneeded keystone CLI calls to get UUIDs for use as parameters on
subsequent CLI calls.  Instead we can just use the user or tenant
name directly and reduce the number of calls made and speed up the
converge a bit.

No change in functionality.

Change-Id: I93deb01d183d9d50098445676c5b49b0b6ae4fea
Closes-Bug: 1479517