Drop token-flush cronjob
We now only support fernet tokens, for which token flushing is no longer
needed, see [1]. So we drop the corresponding cron job and the
attributes to configure it. This also gets rid of the error messages
which it is currently producing.

[1] https://docs.openstack.org/security-guide/identity/tokens.html#fernet-tokens

Change-Id: Ia5fe5f5d0c98ca0a77f03478edfe1777e0a89612
(cherry picked from commit 275c12c3a614df47feac20bb853b55d393b24556)